export API, refactor logger
