+++ lib/abiParser.ts
@@ -2,6 +2,7 @@ import debug from "./debug";
 import { EvmType, VoidType, parseEvmType } from "./typeParser";
 import chalk from "chalk";
 import { MalformedAbiError } from "./errors";
+import { logger } from "./logger";
 
 const { yellow } = chalk;
 
@@ -95,8 +96,7 @@ export function parse(abi: Array<RawAbiDefinition>): Contract {
 
     if (abiPiece.type === "function") {
       if (checkForOverloads(constants, constantFunctions, functions, abiPiece.name)) {
-        // tslint:disable-next-line
-        console.log(yellow(`Detected overloaded constant function ${abiPiece.name} skipping...`));
+        logger.log(yellow(`Detected overloaded constant function ${abiPiece.name} skipping...`));
         return;
       }
 
@@ -113,8 +113,7 @@ export function parse(abi: Array<RawAbiDefinition>): Contract {
     if (abiPiece.type === "event") {
       const eventAbi = (abiPiece as any) as RawEventAbiDefinition;
       if (eventAbi.anonymous) {
-        // tslint:disable-next-line
-        console.log(yellow("Skipping anonymous event..."));
+        logger.log(yellow("Skipping anonymous event..."));
         return;
       }
